fixed change password page for users without password
